requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.7.4
pydantic-settings==2.2.1
email-validator==2.1.0
python-dotenv
redis>=5.0,<6.5
supabase==2.12.0
httpx>=0.27.0,<0.28.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
testcontainers[postgresql,redis]>=4.7.0
respx>=0.21.0
freezegun>=1.5.0
ruff>=0.6.0
mypy>=1.10.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0

# Campaign Execution
sendgrid==6.11.0
twilio==8.13.0

# Additional utilities
python-dateutil==2.8.2

# AI/LLM
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
sentence-transformers>=3.0.0
celery[redis]>=5.4.0
kombu>=5.6.0,<6.0.0
flower>=2.0.0
structlog>=24.0.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-sqlalchemy>=0.46b0
opentelemetry-exporter-otlp>=1.25.0
prometheus-client>=0.20.0